[refactor] 상품홍보페이지 상세 컴포넌트 분리
